refactor: 스키마 마크업 추가 방법 수정 - 기존 `generateMetadata` 방식 확인 결과 스키마 마크업이 `<meta/>` 태그로 추가되어 검색 안되는 오류 발생 - `script` 태그를 이용해 스키마 마크업 추가
